fix: siwf sign up failing when msa already exists (#100)
# Goal
The goal of this PR is to fix a failure when SIWF goes through the sign
up flow and the msa already exists.

# Discussion
- Error occurred because the `BlockchainService` was throwing an error
when the `msaId` was undefined, which resulted in the `account-service`
not returning the proper response in the webhook callback to the
`social-app-template` backend.

# Changes

- [x] `BlockchainService` was refactored so that all events are
processed individually. Therefore, when a `DelegationGranted` event was
found, the proper response is sent to the webhook.