[bug] content-publisher needs to be its own source of truth for announcement attachments  (#341)
# Description
As a user, I should not have to keep track of the association between an
uploaded asset reference ID and its media type. More specifically, I
should not be able to tell content-publisher the type of an asset in an
Announcment publish request (Broadcast, Reply, etc), as that opens up an
attack vector to specify a media type that does not match the actual
uploaded asset. content-publisher should be its own source of truth in
coordinating asset type with the actual asset.

- [x] Remove `type` from `BroadcastDto`, `UpdateDto`, `ReplyDto`
   -  Removed from `AssetDto`, which is included by all of the above
- [x] Update OpenAPI spec
   -  Updated using the `generate` scripts 
- [x] Add `type` to cached metadata for uploaded asset

Closes #338