Refactor BullMQ setup to a common shared component (#565)
This PR takes multiple Gateway 'queues' modules and refactors to a
single shared 'queue' module. Not much else going on here, as it's
pretty much only handling connection setup & monitoring, and was 100%
cloned in each service to begin with.

- Makes `QueueModule` a DynamicModule, with `.forRoot()` to pass in the
queue configs, so that each service can pass in its own. Previously each
service had a carbon copy QueueModule with const queues configured in
the module.