fix: use stock Frequency node image (#663)
Switches from `dsnp/instant-seal-node-with-deployed-schemas` to
`frequencychain/stanadlone-node` for Frequency node Docker image, now
that the standard image includes the Mainnet schemas in the genesis
block. (Should result in a speedier node startup, as we do not need to
wait for the schemas to be deployed by an embedded script)

Also added a health check command to the container so that any services
that depend on it won't come up until the RPC node is ready to serve
requests.