feat: eliminate some chain RPC calls in favor of cached values (#758)
This PR attempts to eliminate a number of RPC calls to the chain from
the critical path of submitting a transaction. By caching certain values
retrieved from the chain, we both increase job throughput in the Gateway
services, and reduce load on the chain RPC node.

- Makes all queue workers' concurrencies configurable via env
`<queueName>_QUEUE_WORKER_CONCURRENCY`
- Manually specify block hash & era for transaction signing and
submission; eliminates anywhere from 3-7 RPC calls to the chain
- Keep current capacity usage and latest nonce updated in Redis cache on
a defined interval. This takes it out of the critical path of submitting
a transaction. (It's OK not to get the absolute latest from the chain
each time)
- If we detect an error submitting a transaction, un-reserve the none
key that we submitted

Related issue: #750

---------

Co-authored-by: Matthew Orris <1466844+mattheworris@users.noreply.github.com>