fix: temporarily bump Redis commandTimeout to 25s (#796)
We ran into a problem where the BullMQ "delayed" queue can cause Bull to
poll with a delay longer than the default 10s timeout for Redis
commands, causing polls to throw "CommandTimeout" errors and delayed
jobs to be stuck in the queue. This PR is a hotfix to increase the
command timeout temporarily until a more correct solution can be
determined.