Content publishing; DELETE /v1/content/:msaId deprecation and new POST /v2/content/:msaId/tombstones (#816)
Closes #756 

I've followed the deprecation pattern that was used in
`/v1/asset/upload` however I was unsuccessful when testing the routes.
Seems I wasn't able to setup my gateway properly locally, will continue
to work on doing that but I'll drop the code here if someone wants to
test it out on their end.

Still left to do:
1. Write tests
2. Generate API docs 

cc: @JoeCap08055

---------

Co-authored-by: Wil Wade <wil.wade@projectliberty.io>
Co-authored-by: Joe Caputo <joseph.caputo@projectliberty.io>