Attempt to improve speed of IPFS publishing by checking for CID existence via gatewayUrl (#821)
# Problem

With a large pinset, `ipfs.pin.ls` can be unacceptably slow, and might
block other IPFS requests.

# Solution

Changes the behavior of IpfsService methods `getPinned` and `getInfo` to
make a HTTP HEAD request via the configured `ipfsGatewayUrl` instead of
calling `ipfs.pin.ls`.

A new method, `checkExists`, was added. It uses the [`Cache-Control:
Only-If-Cached`](https://specs.ipfs.tech/http-gateways/path-gateway/#only-if-cached)
header to ensure the file is not looked for on the wider IPFS network.

The old method, `isPinned`, was not removed, but is currently unused.

## Steps to Verify:

1. Ensure content publishing service is working correctly.
2. Analyze performance.

---------

Co-authored-by: Wil Wade <wil.wade@projectliberty.io>