Announce correct batch file size (#828)
# Problem

`ipfs.dag.stat` gives a `size` value that is not the same as the actual
size of a file when downloaded.
This is because the DAG stores additional info beyond the raw file
bytes.

An example can be seen in the difference between `dag.stat` and
`files.stat`, where the actual length of the file is 87,699 bytes:

`dag.stat` reports the incorrect (for our purposes) value (twice, for
good measure):

```
% curl -s -d '' '$IPFS_RPC_URL/api/v0/dag/stat?arg=bafybeibkvwtaugltazjt7oeq6xahz2qwvd5sm75twsht3e45niuxnfg654' | jq .
{
  "UniqueBlocks": 1,
  "TotalSize": 87713,
  "Ratio": 1,
  "DagStats": [
    {
      "Cid": "bafybeibkvwtaugltazjt7oeq6xahz2qwvd5sm75twsht3e45niuxnfg654",
      "Size": 87713,
      "NumBlocks": 1
    }
  ]
}
```

... whereas `files.stat` gives us both the file size and the DAG size:

```
% curl -s -d '' '$IPFS_RPC_URL/api/v0/files/stat?arg=/ipfs/bafybeibkvwtaugltazjt7oeq6xahz2qwvd5sm75twsht3e45niuxnfg654&progress=true' | jq .
{
  "Hash": "bafybeibkvwtaugltazjt7oeq6xahz2qwvd5sm75twsht3e45niuxnfg654",
  "Size": 87699,
  "CumulativeSize": 87713,
  "Blocks": 0,
  "Type": "file"
}
```

The correct value is also present in a `HEAD` request via the gateway:

```
% curl -s --head '$IPFS_GATEWAY_URL/ipfs/bafybeibkvwtaugltazjt7oeq6xahz2qwvd5sm75twsht3e45niuxnfg654' | grep Content-Length:
Content-Length: 87699
```

# Solution

Since we can get the correct size from the `Content-Length` header of
the `HEAD /ipfs/[CID]` response, and we're already using this as an
existence check, I modified the code to get size info from this request
as well.

It has the nice side effect of one less request to the IPFS server.

The `ipfsService.getInfo(cid)` method has been removed since it was
otherwise unused, and was not providing the output expected. The batch
announcer has been updated to get the file size value directly via the
new method, `ipfsService.contentLengthInLocalGateway(cid)`.

## Steps to Verify:

1. Check the `payload_length` reported in a Frequency IpfsMessage
against the actual file size of the batch.