fix: do not return a Promise inside try...catch (#916)
# Problem
A missing `await` was causing errors to be mis-handled.

# Details
Normally, in an `async` Javascript method, it is not necessary to
'await' a value that is being returned; the logic being that since an
async function wraps its return value in a Promise anyway, returning a
Promise directly is transparent to the caller. So, the following is
perfectly fine:
```javascript
async function foo(): Promise<number> {
  // do some async stuff
  return value;
}

async function bar(): Promise<number> {
  return foo();
}
```

In the above, 'bar' doesn't need to 'await foo()', because it will
return `Promise<number>` regardless.

However, if we modify the code as follows:
```javascript
async function foo(): Promise<number> {
  // do some async stuff
  if (has_error) {
    throw new Error('I screwed up');
  }

  return value;
}

async bar(): Promise<number> {
   try {
    return foo();
  } catch (err: any) {
    throw Error('foo screwed up');
  }
}
```

In the above, the statement `return foo();` will **always** return a
`Promise<number>`. If 'foo' throws, any callers of 'bar', if they
'await' or '.catch(...)' the return value, will only ever get the error
'I screwed up'. The error 'foo screwed up' will **never** be thrown,
because 'foo' is always a Promise that hasn't resolved or rejected yet
at the time 'bar' returns.

So, the moral of the story is: it's safe to return an encapsulated async
function's value directly inside another async function _unless the call
is inside a try...catch block_!