Add missing delay for exponential backoff (#956)
# Problem Exponential backoff had no delay configured so operations that
encountered errors would be retried immediately.

# Solution

Introduce a reasonable delay.

I used 1000ms as the default, but 6000ms for the Publish Queue, where
errors have been observed, and waiting at least until the next block
might allow a working nonce to be retrieved.

It is likely unnecessary to add a delay where `attempts` is set to `1`,
but I added those as a best practice to avoid confusion if `attempts` is
ever changed.

No tests included as I think we'd just be testing whether the queueing
library worked.