Remove ultra light CI load test on PRs (#962)
# Problem

Currently, the CI is having issues meeting load test thresholds. This is
variable as we are not sure what kind of CPU & memory allocation CI is
running on.

# Solution

For the time being, instead of arbitrarily lowering thresholds, lets
remove this CI job, and research how to make reasonable load tests a
part of CI. Will create a related issue.